ipc: move IPC specific component logic into IPC helper.
comp_new() and dependencies are all specific to the IPC ABI MAJOR,
move them to IPC helper.

No code changes, move only.
