+++ src/ipc/helper-ipc3.c
@@ -10,6 +10,7 @@
 #include <sof/audio/pipeline.h>
 #include <sof/common.h>
 #include <sof/drivers/idc.h>
+#include <sof/drivers/interrupt.h>
 #include <sof/ipc/topology.h>
 #include <sof/ipc/common.h>
 #include <sof/ipc/msg.h>
@@ -31,6 +32,8 @@
 #include <stdbool.h>
 #include <stddef.h>
 #include <stdint.h>
+
+extern struct tr_ctx comp_tr;
 
 void ipc_build_stream_posn(struct sof_ipc_stream_posn *posn, uint32_t type,
 			   uint32_t id)
@@ -55,6 +58,222 @@ void ipc_build_trace_posn(struct sof_ipc_dma_trace_posn *posn)
 	posn->rhdr.hdr.cmd =  SOF_IPC_GLB_TRACE_MSG |
 		SOF_IPC_TRACE_DMA_POSITION;
 	posn->rhdr.hdr.size = sizeof(*posn);
+}
+
+/* Function overwrites PCM parameters (frame_fmt, buffer_fmt, channels, rate)
+ * with buffer parameters when specific flag is set.
+ */
+static void comp_update_params(uint32_t flag,
+			       struct sof_ipc_stream_params *params,
+			       struct comp_buffer *buffer)
+{
+	if (flag & BUFF_PARAMS_FRAME_FMT)
+		params->frame_fmt = buffer->stream.frame_fmt;
+
+	if (flag & BUFF_PARAMS_BUFFER_FMT)
+		params->buffer_fmt = buffer->buffer_fmt;
+
+	if (flag & BUFF_PARAMS_CHANNELS)
+		params->channels = buffer->stream.channels;
+
+	if (flag & BUFF_PARAMS_RATE)
+		params->rate = buffer->stream.rate;
+}
+
+int comp_verify_params(struct comp_dev *dev, uint32_t flag,
+		       struct sof_ipc_stream_params *params)
+{
+	struct list_item *buffer_list;
+	struct list_item *source_list;
+	struct list_item *sink_list;
+	struct list_item *clist;
+	struct list_item *curr;
+	struct comp_buffer *sinkb;
+	struct comp_buffer *buf;
+	int dir = dev->direction;
+	uint32_t flags = 0;
+
+	if (!params) {
+		comp_err(dev, "comp_verify_params(): !params");
+		return -EINVAL;
+	}
+
+	source_list = comp_buffer_list(dev, PPL_DIR_UPSTREAM);
+	sink_list = comp_buffer_list(dev, PPL_DIR_DOWNSTREAM);
+
+	/* searching for endpoint component e.g. HOST, DETECT_TEST, which
+	 * has only one sink or one source buffer.
+	 */
+	if (list_is_empty(source_list) != list_is_empty(sink_list)) {
+		if (!list_is_empty(source_list))
+			buf = list_first_item(&dev->bsource_list,
+					      struct comp_buffer,
+					      sink_list);
+		else
+			buf = list_first_item(&dev->bsink_list,
+					      struct comp_buffer,
+					      source_list);
+
+		buffer_lock(buf, &flags);
+
+		/* update specific pcm parameter with buffer parameter if
+		 * specific flag is set.
+		 */
+		comp_update_params(flag, params, buf);
+
+		/* overwrite buffer parameters with modified pcm
+		 * parameters
+		 */
+		buffer_set_params(buf, params, BUFFER_UPDATE_FORCE);
+
+		/* set component period frames */
+		component_set_period_frames(dev, buf->stream.rate);
+
+		buffer_unlock(buf, flags);
+	} else {
+		/* for other components we iterate over all downstream buffers
+		 * (for playback) or upstream buffers (for capture).
+		 */
+		buffer_list = comp_buffer_list(dev, dir);
+		clist = buffer_list->next;
+
+		while (clist != buffer_list) {
+			curr = clist;
+
+			buf = buffer_from_list(curr, struct comp_buffer, dir);
+
+			buffer_lock(buf, &flags);
+
+			clist = clist->next;
+
+			comp_update_params(flag, params, buf);
+
+			buffer_set_params(buf, params, BUFFER_UPDATE_FORCE);
+
+			buffer_unlock(buf, flags);
+		}
+
+		/* fetch sink buffer in order to calculate period frames */
+		sinkb = list_first_item(&dev->bsink_list, struct comp_buffer,
+					source_list);
+
+		buffer_lock(sinkb, &flags);
+
+		component_set_period_frames(dev, sinkb->stream.rate);
+
+		buffer_unlock(sinkb, flags);
+	}
+
+	return 0;
+}
+
+static const struct comp_driver *get_drv(struct sof_ipc_comp *comp)
+{
+	struct comp_driver_list *drivers = comp_drivers_get();
+	struct list_item *clist;
+	const struct comp_driver *drv = NULL;
+	struct comp_driver_info *info;
+	struct sof_ipc_comp_ext *comp_ext;
+	uint32_t flags;
+
+	irq_local_disable(flags);
+
+	/* do we have extended data ? */
+	if (!comp->ext_data_length)
+		goto comp_type_match;
+
+	/* Basic sanity check of the total size and extended data
+	 * length. A bit lax because in this generic code we don't know
+	 * which derived comp we have and how much its specific members
+	 * add.
+	 */
+	if (comp->hdr.size < sizeof(*comp) + comp->ext_data_length) {
+		tr_err(&comp_tr, "Invalid size, hdr.size=0x%x, ext_data_length=0x%x\n",
+		       comp->hdr.size, comp->ext_data_length);
+		goto out;
+	}
+
+	comp_ext = (struct sof_ipc_comp_ext *)
+		   ((uint8_t *)comp + comp->hdr.size -
+		    comp->ext_data_length);
+
+	/* UUID is first item in extended data - check its big enough */
+	if (comp->ext_data_length < UUID_SIZE) {
+		tr_err(&comp_tr, "UUID is invalid!\n");
+		goto out;
+	}
+
+	/* search driver list with UUID */
+	list_for_item(clist, &drivers->list) {
+		info = container_of(clist, struct comp_driver_info,
+				    list);
+		if (!memcmp(info->drv->uid, comp_ext->uuid,
+			    UUID_SIZE)) {
+			tr_dbg(&comp_tr,
+			       "get_drv_from_uuid(), found driver type %d, uuid %pU",
+			       info->drv->type,
+			       info->drv->tctx->uuid_p);
+			drv = info->drv;
+			goto out;
+		}
+	}
+
+	tr_err(&comp_tr, "get_drv(): the provided UUID (%8x%8x%8x%8x) doesn't match to any driver!",
+	       *(uint32_t *)(&comp_ext->uuid[0]),
+	       *(uint32_t *)(&comp_ext->uuid[4]),
+	       *(uint32_t *)(&comp_ext->uuid[8]),
+	       *(uint32_t *)(&comp_ext->uuid[12]));
+
+	goto out;
+
+comp_type_match:
+	/* search driver list for driver type */
+	list_for_item(clist, &drivers->list) {
+		info = container_of(clist, struct comp_driver_info, list);
+		if (info->drv->type == comp->type) {
+			drv = info->drv;
+			goto out;
+		}
+	}
+
+out:
+	irq_local_enable(flags);
+	return drv;
+}
+
+struct comp_dev *comp_new(struct sof_ipc_comp *comp)
+{
+	struct comp_dev *cdev;
+	const struct comp_driver *drv;
+
+	/* find the driver for our new component */
+	drv = get_drv(comp);
+	if (!drv) {
+		tr_err(&comp_tr, "comp_new(): driver not found, comp->type = %u",
+		       comp->type);
+		return NULL;
+	}
+
+	/* validate size of ipc config */
+	if (IPC_IS_SIZE_INVALID(*comp_config(comp))) {
+		IPC_SIZE_ERROR_TRACE(&comp_tr, *comp_config(comp));
+		return NULL;
+	}
+
+	tr_info(&comp_tr, "comp new %pU type %d id %d.%d",
+		drv->tctx->uuid_p, comp->type, comp->pipeline_id, comp->id);
+
+	/* create the new component */
+	cdev = drv->ops.create(drv, comp);
+	if (!cdev) {
+		comp_cl_err(drv, "comp_new(): unable to create the new component");
+		return NULL;
+	}
+
+	list_init(&cdev->bsource_list);
+	list_init(&cdev->bsink_list);
+
+	return cdev;
 }
 
 int ipc_pipeline_new(struct ipc *ipc,
